refactor api graph dto types
